requirements.txt
fastapi
uvicorn

# async_scrape.py
requests
bs4
colorama
httpx

# mode_manual.py
python-dateutil
